+++ src/main/java/by/itacademy/javaenterprise/lepnikau/app/entity/Parent.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class Parent{
-    private int id;
-    private int studentId;
+    private long id;
+    private long studentId;
     private String lastName;
     private String firstName;
     private String patronymic;
